Still stuck
Assertion `cur_target >= 0 && cur_target < n_classes' failed.

Having trouble tracking down the cause of this but I think it may have to do with how I'm using an embedding layer with an lstm layer.  Pretty sure this means that the output of the forward pass is outside of the dictionary range.

In an effort to solve this I managed to track down two big bugs.  I was setting the model to zero grad instead of the optimizer, and I was indexing my conversion dictionaries from one like in a previous exercise, but we're not padding last batches with zeroes like in that exercise.