Training Takes Too Long
I finally got it to train by fixing a bunch of issues with the hyperparameters.  The biggest issue was the output_size, which needed to be set to the size of the entire vocabulary.

With a retained graph for the backward step I was unable to get the model to complete even one epoch while burning through about four hours of GPU time.  I need to figure out how to fix the issue without turning on retain_graph in the forward_back_prop.