Still stuck
I'm trying to conserve GPU time now.  I've been able to get it to train almost to below a loss of five, but the model gets bogged down by accumulated past weights by the time it gets about halfway through the script (1 epoch).  I've had to turn on retain_graph on the training, so not I'm trying to figure out how to use .detach().  I'm unable to call the function on the weights because they are a tuple.  I can call it on the loss itself, but I'm not sure that actually does anything.  I'm torn between writing an in-line function that goes through each value in the weights or just figuring out a way to call init_weights inside the forward_back_prop function.  This feels like a stupid place to get stuck but everything I find on .detach is using in a way that doesn't quite work in this situation.  Gotta keep working on this.