It works, kind of
I managed to train my model below 3.5 loss (the goal) on my home machine, but there were compatibility issues with some of the libraries (I think).  So, now the code can't generate new scripts properly.  The trick was to increase the batch sizes and the number of epochs.  My personal machine was able to handle that easily.